pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blq-cli"
version = "0.2.1"
description = "Buidl Log Query - capture and query build/test logs with DuckDB"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Teague Sterling", email = "teaguesterling@gmail.com" }
]
keywords = ["duckdb", "logs", "build", "test", "ci", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "duckdb>=1.4.3",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov",
    "ruff",
    "mypy",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]
mcp = [
    "fastmcp>=2.0.0",
]

[project.scripts]
blq = "blq.cli:main"

[project.urls]
Homepage = "https://blq-cli.readthedocs.com/"
Repository = "https://github.com/teaguesterling/blq-cli"
Issues = "https://github.com/teaguesterling/blq-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/blq"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
